fixed deployment
