+++ src/app/dashboard/page.tsx
@@ -1,0 +1,13 @@
+export default function Page() {
+  return (
+    <div>
+      <h1>You are on a Dashboard page!</h1>
+      <p>
+        Lorem ipsum dolor sit, amet consectetur adipisicing elit. Voluptas,
+        laudantium voluptatibus. Dolorem, dolore earum quis corrupti voluptatum
+        delectus esse, harum aliquam ex quo adipisci culpa minima qui suscipit
+        autem velit.
+      </p>
+    </div>
+  );
+}
