+++ src/app/page.tsx
@@ -1,0 +1,5 @@
+import Link from "next/link";
+
+export default function Page() {
+  return <Link href="/dashboard">Dashboard page</Link>;
+}
